pyproject.toml
[tool.poetry]
name = "tools"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [
    { include = "tools" }
]

[tool.poetry.dependencies]
python = ">=3.8, <4"
PyGithub = "^1.58.2"
requests = "^2.31.0"
duckduckgo-search = "^3.1.1"
openai = "^0.27.7"
tiktoken = "^0.4.0"
pyyaml = "^6.0.1"
pylanguagetool = "^0.9.2"
markdown2 = "^2.4.10"
bs4 = "^0.0.1"

[tool.poetry.scripts]
fact-check = "tools.fact_checker:main"
payout-calc = "tools.payout_calc:main"
grammar-check =  "tools.grammar_checker:main"
duplication-check = "tools.duplication_checker:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
